% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_pkg_sql.R
\name{exec_pkg_sql}
\alias{exec_pkg_sql}
\title{Execute a packaged SQL file (lives in inst/sql of this package) using utHelpR}
\usage{
exec_pkg_sql(name, dsn = "edify")
}
\arguments{
\item{name}{SQL filename (with or without .sql)}

\item{dsn}{DSN passed through to utHelpR::get_data_from_sql_file}
}
\value{
Tibble/data.frame
}
\description{
Execute a packaged SQL file (lives in inst/sql of this package) using utHelpR
}
\examples{
# exec_pkg_sql("kpi_rollup")
}
