% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_data_importer.R
\name{file_importer}
\alias{file_importer}
\title{Create an object for importing R objects from files}
\usage{
file_importer(path)
}
\arguments{
\item{path}{Scalar character. The filepath to the data to be imported.}
}
\value{
A `file_importer` object (parent: `data_importer`) to be used when importing data.
}
\description{
Create an object for importing R objects from files
}
