% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_current_dtcs}
\alias{get_current_dtcs}
\title{get_current_dtcs}
\usage{
get_current_dtcs()
}
\value{
An integer
}
\description{
This function returns the current days to class start dtcs

The days to class start is a measure of how many days before the start of classes
an events fall on. The first day of class has a dtcs of 0. The days are negative
prior to the start of the semester, and positive after the start of the semester.
}
