% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_data_from_pin}
\alias{get_data_from_pin}
\title{get_data_from_pin}
\usage{
get_data_from_pin()
}
\value{
Returns a data frame from the daily enrollment pin.
}
\description{
This function pulls data from the daily enrollment pin.
}
