% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_housing_value_box_data.R
\name{get_housing_value_box_data}
\alias{get_housing_value_box_data}
\title{Convenience wrapper to produce housing value box data}
\usage{
get_housing_value_box_data(dsn = NULL, context = NULL)
}
\arguments{
\item{dsn}{Optional data source name / connection (if used).}

\item{context}{Optional context/environment object.}
}
\value{
Tibble of value box metrics.
}
\description{
Convenience wrapper to produce housing value box data
}
