% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_connect.R
\name{get_pins_board}
\alias{get_pins_board}
\title{Access the pins board that contains data for this app}
\usage{
get_pins_board(
  config_file = NULL,
  server_key = "connect_server",
  account_key = "connect_account",
  api_key_key = "connect_api_key"
)
}
\value{
A pins board object
}
\description{
Access the pins board that contains data for this app
}
