% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_functions.R
\name{get_val_box_color_generic}
\alias{get_val_box_color_generic}
\title{get_val_box_color_generic}
\usage{
get_val_box_color_generic(data, var_name, threshold = 0)
}
\arguments{
\item{data}{a data frame with a single row}

\item{var_name}{the name of the variable to use}

\item{threshold}{the threshold value to compare against}
}
\value{
a theme color
}
\description{
get_val_box_color_generic
}
