% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_functions.R
\name{get_value_box}
\alias{get_value_box}
\title{get_value_box}
\usage{
get_value_box(data, metric, lookup)
}
\arguments{
\item{data}{a data frame with a single row}

\item{metric}{the metric to display}

\item{lookup}{a data frame with two columns: metric and test}
}
\value{
A value box
}
\description{
get_value_box
}
