% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GolemKpiSnapshotSuite-package.R
\docType{package}
\name{GolemKpiSnapshotSuite-package}
\alias{GolemKpiSnapshotSuite}
\alias{GolemKpiSnapshotSuite-package}
\title{GolemKpiSnapshotSuite: Shared Golem Modules and KPI Snapshot Utilities}
\description{
Shared modules, utilities, and templates for mobile cabinet KPI Shiny apps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ut-effectiveness/GolemKpiSnapshotSuite}
  \item Report bugs at \url{https://github.com/ut-effectiveness/GolemKpiSnapshotSuite/issues}
}

}
\author{
\strong{Maintainer}: Craig Demke \email{craig.demke@utahtech.edu}

}
\keyword{internal}
