% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_build.R
\name{housing_build}
\alias{housing_build}
\title{Build housing derived metrics}
\usage{
housing_build(raw)
}
\arguments{
\item{raw}{Raw housing dataset (tibble/data.frame).}
}
\value{
Tibble with derived metrics.
}
\description{
Build housing derived metrics
}
