% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_housing.R, R/housing_detailed_table.R
\name{housing_detailed_table}
\alias{housing_detailed_table}
\title{Convenience: return detailed table}
\usage{
housing_detailed_table()

housing_detailed_table()
}
\arguments{
\item{dsn}{Optional data source name.}

\item{context}{Optional context object.}
}
\value{
A tibble / data.frame for display.
}
\description{
Convenience: return detailed table

Housing detailed table
}
