% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_housing.R
\name{housing_value_box_data}
\alias{housing_value_box_data}
\title{Convenience: return just the value box dataset}
\usage{
housing_value_box_data(dsn = "edify", context = "shiny")
}
\description{
Convenience: return just the value box dataset
}
