% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_data_importer.R
\name{import}
\alias{import}
\title{Import an object from a file or pin, or from a SQL query}
\usage{
import(x, ...)
}
\arguments{
\item{x}{A `data_importer` object}

\item{...}{Further arguments.}
}
\value{
The imported object (typically a data-frame)
}
\description{
Import an object from a file or pin, or from a SQL query
}
