% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_connect.R
\name{is_connect}
\alias{is_connect}
\title{Is the app running on Connect?}
\usage{
is_connect()
}
\value{
Boolean. TRUE if the app is running on a Connect server.
}
\description{
Is the app running on Connect?
}
