% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{kpi_branding}
\alias{kpi_branding}
\title{Shared KPI branding assets}
\usage{
kpi_branding(prefix = "gkss", include_css = TRUE)
}
\arguments{
\item{prefix}{resource path prefix}

\item{include_css}{include custom.css & litera_style.css}
}
\description{
Shared KPI branding assets
}
