% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R, R/kpi_logo.R
\name{kpi_logo}
\alias{kpi_logo}
\title{Logo tag}
\usage{
kpi_logo(height = "48px", prefix = "gkss")

kpi_logo(height = "48px", prefix = "gkss")
}
\description{
Logo tag

Simple logo (ut.png) from parent assets
}
