% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kpi_snapshot_ui.R
\name{kpi_snapshot_inner_ui}
\alias{kpi_snapshot_inner_ui}
\title{KPI Snapshot UI Module}
\usage{
kpi_snapshot_inner_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
KPI Snapshot UI Module
}
