% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{make_current_data}
\alias{make_current_data}
\title{make_current_data}
\usage{
make_current_data()
}
\value{
Returns a data frame from the current term of interest and the previous term.
}
\description{
This function pulls the latest term, and the previous terms data

The current term of interest is the term we are watching the headcount for. The
current term is the term the university is currently in. The previous term of
interest is the term we want to compare the current term of interest to.
}
