% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_headcount.R
\name{mod_headcount_server}
\alias{mod_headcount_server}
\title{Headcount module server}
\usage{
mod_headcount_server(
  id,
  device_type = "desktop",
  value_box_data,
  plot_data,
  custom_server = NULL
)
}
\arguments{
\item{id}{Module id.}

\item{device_type}{Reactive or value describing device type.}

\item{value_box_data}{Reactive providing value box data.}

\item{plot_data}{Reactive providing plotting data.}

\item{custom_server}{Optional function to override server internals.}
}
\description{
Headcount module server
}
