% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ipeds.R
\name{mod_ipeds_server}
\alias{mod_ipeds_server}
\title{IPEDS module server}
\usage{
mod_ipeds_server(
  id,
  device_type,
  value_box_data,
  plot_data,
  custom_server = NULL
)
}
\arguments{
\item{id}{Module id.}

\item{device_type}{Reactive or value describing device type.}

\item{value_box_data}{Reactive providing value box data.}

\item{plot_data}{Reactive providing plotting data.}

\item{custom_server}{Optional override (unused here).}
}
\description{
IPEDS module server
}
