% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kpi_snapshot_server.R
\name{mod_kpi_snapshot_server}
\alias{mod_kpi_snapshot_server}
\title{KPI Snapshot Server Module}
\usage{
mod_kpi_snapshot_server(id, custom_server = NULL)
}
\description{
KPI Snapshot Server Module
}
