% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kpi_snapshot_ui.R
\name{mod_kpi_snapshot_ui}
\alias{mod_kpi_snapshot_ui}
\title{mod_kpi_snapshot_ui}
\usage{
mod_kpi_snapshot_ui(id, custom_ui = NULL)
}
\arguments{
\item{id}{module id}
}
\description{
mod_kpi_snapshot_ui
}
