% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_data_importer.R
\name{pin_importer}
\alias{pin_importer}
\title{Create an object for importing R objects from pin boards}
\usage{
pin_importer(board, name)
}
\arguments{
\item{board}{The pin board.}

\item{name}{The name of the pin on the pin board.}
}
\value{
A `pin_importer` object (parent: `data_importer`) to be used when importing data from
  the pin.
}
\description{
Create an object for importing R objects from pin boards
}
