% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{set_default_child_pkg}
\alias{set_default_child_pkg}
\title{Set default child package for config lookups}
\usage{
set_default_child_pkg(pkg)
}
\arguments{
\item{pkg}{child package name}
}
\description{
Set default child package for config lookups
}
