% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_logo.R
\name{title_logo}
\alias{title_logo}
\title{Title logo UI fragment}
\usage{
title_logo(height = "40px", class = NULL)
}
\arguments{
\item{height}{CSS height (character).}

\item{class}{Additional CSS classes.}
}
\value{
A shiny tag.
}
\description{
Title logo UI fragment
}
