% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{Detect client device info (robust)}
\usage{
device()
}
\value{
list(device = <chr>, ... possible extra fields)
}
\description{
Returns a list with at least element `device` (lowercase: mobile|tablet|desktop|unknown).
Safe even if shinybrowser not ready yet.
}
