% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Universal golem config getter for parent and child apps}
\usage{
get_golem_config(
  value,
  package = "GolemKpiSnapshotSuite",
  config = Sys.getenv("GOLEM_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE
)
}
\arguments{
\item{value}{Config value to retrieve}

\item{package}{Name of the package to read config from. Defaults to parent package ("GolemKpiSnapshotSuite").}

\item{config}{Active config to use}

\item{use_parent}{Whether to use parent config logic}
}
\value{
The config value

The config value
}
\description{
Universal golem config getter for parent and child apps
}
