% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Read a config value from a child golem-config.yml}
\usage{
get_golem_config(
  value,
  pkg,
  config = Sys.getenv("GOLEM_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE
)
}
\description{
Read a config value from a child golem-config.yml
}
