% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_golem_config}
\alias{get_golem_config}
\title{Universal golem config getter (shared)}
\usage{
get_golem_config(
  value,
  pkg = "GolemKpiSnapshotSuite",
  config = Sys.getenv("GOLEM_CONFIG_ACTIVE", Sys.getenv("R_CONFIG_ACTIVE", "default")),
  use_parent = TRUE
)
}
\arguments{
\item{value}{Config key}

\item{pkg}{Package that holds the golem-config.yml (default = this package)}

\item{config}{Active config name}

\item{use_parent}{Forwarded to config::get}
}
\description{
Universal golem config getter (shared)
}
