% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_functions.R
\name{get_val_box_goal_generic}
\alias{get_val_box_goal_generic}
\title{Title}
\usage{
get_val_box_goal_generic(
  data,
  var_name,
  threshold = 0,
  neg = " still needed ",
  pos = " ahead "
)
}
\arguments{
\item{data}{a data frame with a single row}

\item{var_name}{the name of the variable to use}

\item{threshold}{the threshold value to compare against}

\item{neg}{a string to display if the value is above the threshold}

\item{pos}{a string to display if the value is below the threshold}
}
\description{
Title
}
