% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_housing.R
\name{housing_build}
\alias{housing_build}
\title{Build housing datasets (value boxes + detailed table)}
\usage{
housing_build(raw)
}
\description{
Returns a list with:
  final_housing_for_valueboxs : value box friendly tibble
  detailed_table_for_DT       : detailed building-level tibble
}
