% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_detailed_table.R
\name{housing_detailed_table}
\alias{housing_detailed_table}
\title{Housing detailed table}
\usage{
housing_detailed_table()
}
\value{
A tibble / data.frame for display.
}
\description{
Housing detailed table
}
