% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_housing.R
\name{housing_detailed_table}
\alias{housing_detailed_table}
\title{Convenience: return detailed table}
\usage{
housing_detailed_table(dsn = "edify", context = "shiny")
}
\description{
Convenience: return detailed table
}
