% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_pull_raw.R
\name{housing_pull_raw}
\alias{housing_pull_raw}
\title{Pull raw housing data}
\usage{
housing_pull_raw()
}
\arguments{
\item{dsn}{Optional data source name.}

\item{context}{Optional context object.}
}
\value{
Raw housing data (tibble).
}
\description{
Pull raw housing data
}
