% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_housing.R, R/housing_value_box_data.R
\name{housing_value_box_data}
\alias{housing_value_box_data}
\title{Convenience: return just the value box dataset}
\usage{
housing_value_box_data(dsn = NULL, context = NULL)

housing_value_box_data(dsn = NULL, context = NULL)
}
\arguments{
\item{dsn}{Optional data source name / connection (if used).}

\item{context}{Optional context/environment object.}
}
\value{
Tibble ready for value box creation.
}
\description{
Convenience: return just the value box dataset

Build housing value box data
}
