% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_value_box_data.R
\name{housing_value_box_data}
\alias{housing_value_box_data}
\title{Build housing value box data}
\usage{
housing_value_box_data()
}
\value{
Tibble ready for value box creation.
}
\description{
Build housing value box data
}
