% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{kpi_branding}
\alias{kpi_branding}
\title{Register parent asset path + (optionally) navbar CSS only}
\usage{
kpi_branding(prefix = "gkss", include_navbar = TRUE)
}
\arguments{
\item{prefix}{resource path prefix}

\item{include_navbar}{logical include navbar.css}
}
\description{
Register parent asset path + (optionally) navbar CSS only
}
