% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{kpi_branding}
\alias{kpi_branding}
\title{Register parent assets & inject full mobileCabinetKpi styling}
\usage{
kpi_branding(prefix = "gkss", include_css = TRUE)
}
\arguments{
\item{prefix}{resource path prefix}

\item{include_css}{logical}
}
\description{
Register parent assets & inject full mobileCabinetKpi styling
}
