% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{kpi_logo}
\alias{kpi_logo}
\title{Logo tag}
\usage{
kpi_logo(file = "ie_logo.png", prefix = "gkss", ...)
}
\arguments{
\item{file}{image filename in inst/app/www}

\item{prefix}{resource prefix}
}
\description{
Logo tag
}
