% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R, R/litera_theme.R
\name{litera_theme}
\alias{litera_theme}
\title{Simple Litera theme (system fonts only)}
\usage{
litera_theme()

litera_theme()
}
\description{
Simple Litera theme (system fonts only)

Shared Litera theme (same as mobileCabinetKpi)
}
