% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R
\name{litera_theme}
\alias{litera_theme}
\title{Litera Bootstrap Theme for Shiny UI}
\usage{
litera_theme()
}
\value{
A bslib theme object using the Litera Bootswatch theme.
}
\description{
Litera Bootstrap Theme for Shiny UI
}
