% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_housing.R
\name{mod_housing_server}
\alias{mod_housing_server}
\title{housing_tab server Function}
\usage{
mod_housing_server(
  id,
  device_type = "Desktop",
  value_box_data,
  dt_data,
  comparison_mode
)
}
\arguments{
\item{id}{Module id.}

\item{device_type}{Reactive or value describing device type.}

\item{value_box_data}{Reactive providing value box data.}

\item{dt_data}{Reactive providing plotting data.}

\item{custom_server}{Optional function to override server internals.}
}
\description{
housing_tab server Function
}
