% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_student_type.R
\name{mod_student_type_server}
\alias{mod_student_type_server}
\title{main_tab Server Functions}
\usage{
mod_student_type_server(id, device_type = "Desktop", value_box_data, plot_data)
}
\arguments{
\item{id}{Module id.}

\item{device_type}{Reactive or value describing device type.}

\item{value_box_data}{Reactive providing value box data.}

\item{plot_data}{Reactive providing plotting data.}

\item{custom_server}{Optional function to override server internals.}
}
\description{
main_tab Server Functions
}
