% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R
\name{title_logo}
\alias{title_logo}
\title{style_utils}
\usage{
title_logo(
  text = getOption("gkss.app_title", "Enrollment KPI"),
  prefix = "gkss",
  logo_src = "ut.png",
  width = 80,
  height = 40,
  show_logo = TRUE
)
}
\arguments{
\item{text}{Title text}

\item{prefix}{Resource path prefix (from kpi_branding)}

\item{logo_src}{Logo file name in inst/app/www}

\item{width, height}{Logo size (px)}

\item{show_logo}{Show logo?}
}
\description{
Title logo component (mobile layout mimic)
}
\details{
Mobile KPI style top bar (logo left, title right)
}
