% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R
\name{title_logo}
\alias{title_logo}
\title{style_utils}
\usage{
title_logo(
  text = getOption("gkss.app_title", "Enrollment KPI"),
  logo = TRUE,
  prefix = "gkss",
  logo_src = "ut.png",
  width = 80,
  height = 40
)
}
\arguments{
\item{text}{Title string (defaults to option gkss.app_title or "Enrollment KPI")}

\item{logo}{Include logo image}

\item{prefix}{Resource path prefix added by kpi_branding()}

\item{logo_src}{Image filename in inst/app/www}

\item{width, height}{Logo dimensions (px)}

\item{class}{Wrapper div class}

\item{text_class}{CSS class for title text}
}
\description{
Create title logo
}
\details{
Returns UT logo with correct dimensions for app title
}
