% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R, R/title_logo.R
\name{title_logo}
\alias{title_logo}
\title{style_utils}
\usage{
title_logo(
  file = "ie_logo.png",
  height = "40px",
  alt = "Logo",
  prefix = "gkss"
)

title_logo(
  file = "ie_logo.png",
  height = "40px",
  alt = "Logo",
  prefix = "gkss"
)
}
\arguments{
\item{file}{image filename inside inst/app/www}

\item{height}{CSS height (character).}

\item{alt}{alt text}

\item{prefix}{resource path prefix (auto-added if needed)}
}
\description{
Create title logo
}
\details{
Returns UT logo with correct dimensions for app title.
}
