% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R
\name{title_logo}
\alias{title_logo}
\title{style_utils}
\usage{
title_logo(
  text = getOption("gkss.app_title", "Enrollment KPI"),
  prefix = "gkss",
  logo_src = "ut.png",
  width = 80,
  height = 40,
  text_size = "22pt",
  text_right = 80,
  text_top = 0,
  text_margin_top = 10,
  container_width = 150
)
}
\description{
Title logo component (mobile layout mimic)
}
