% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_style_utils.R
\name{title_logo}
\alias{title_logo}
\title{style_utils}
\usage{
title_logo(
  text = getOption("gkss.app_title", "Enrollment KPI"),
  logo = TRUE,
  prefix = "gkss",
  logo_src = "ut.png",
  width = 80,
  height = 40
)
}
\arguments{
\item{text}{Title text.}

\item{logo}{Show logo (TRUE/FALSE).}

\item{prefix}{Resource path prefix registered by kpi_branding().}

\item{logo_src}{Logo filename under inst/app/www.}

\item{width}{Logo width}

\item{hight}{Logo height}
}
\description{
Title logo component (mobile layout mimic)
}
