% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_logo.R
\name{title_logo}
\alias{title_logo}
\title{Title logo UI fragment}
\usage{
title_logo(
  file = "ie_logo.png",
  height = "40px",
  alt = "Logo",
  prefix = "gkss"
)
}
\arguments{
\item{file}{image filename inside inst/app/www}

\item{height}{CSS height (character).}

\item{alt}{alt text}

\item{prefix}{resource path prefix (auto-added if needed)}
}
\description{
Title logo UI fragment
}
